test: consolidate no-unused-vars cases
